More login changes
