+++ Sluggo-SwiftUI/Views/Components/IconSluggo.swift
@@ -1,0 +1,18 @@
+//
+//  IconSluggo.swift
+//  IconSluggo
+//
+//  Created by Andrew Gavgavian on 9/3/21.
+//
+
+import SwiftUI
+
+struct IconSluggo: View {
+    
+    var body: some View {
+        Image("IconSluggo")
+            .resizable()
+            .scaledToFit()
+    }
+    
+}
