+++ Sluggo-SwiftUI/Views/LoginView.swift
@@ -22,12 +22,9 @@ struct LoginView: View {
     
     
     var body: some View {
-        Spacer()
-        
         Form {
-            Image(uiImage: UIImage(named: "AppIcon") ?? UIImage())
-                .resizable()
-                .scaledToFit()
+            IconSluggo()
+                .padding()
             
             Text("Login Page")
                 .padding()
@@ -50,7 +47,7 @@ struct LoginView: View {
                 .padding()
             
             HStack() {
-                CheckBoxView(checked: $isPersistance, caption: "Remember Me?")
+                CheckBox(checked: $isPersistance, caption: "Remember Me?")
                     .padding()
                 
                 Spacer()
@@ -59,6 +56,7 @@ struct LoginView: View {
                     self.attemptLogin()
                 }
                 .buttonStyle(.borderedProminent)
+                .controlSize(.large)
             }
             
             
@@ -115,47 +113,5 @@ struct LoginView: View {
 struct LoginView_Previews: PreviewProvider {
     static var previews: some View {
         LoginView(showModal: .constant(true))
-    }
-}
-
-struct CheckBoxView: View {
-    @Binding var checked: Bool
-    var caption: String
-    
-    var body: some View {
-        HStack {
-            Image(systemName: checked ? "checkmark.square.fill" : "square")
-                .foregroundColor(checked ? Color(UIColor.systemBlue) : Color.secondary)
-                .onTapGesture {
-                    self.checked.toggle()
-                }
-            
-            Text(caption)
-                .font(.caption)
-                .fontWeight(.regular)
-        }
-    }
-}
-
-
-
-struct LoginTextField: View {
-    @State var placeholder: String
-    @Binding var textValue: String
-    @State var textStyle: UITextContentType
-    var isSecure: Bool = false
-    var body: some View {
-        if isSecure == true {
-            SecureField(placeholder, text: $textValue)
-                .textContentType(textStyle)
-                .autocapitalization(.none)
-                .disableAutocorrection(true)
-        } else {
-            TextField(placeholder, text: $textValue)
-                .textContentType(textStyle)
-                .autocapitalization(.none)
-                .disableAutocorrection(true)
-        }
-        
     }
 }
